Added stub for notifications processing

+++ src/main/java/com/prefserver/dao/PlayerDao.java
@@ -14,5 +14,5 @@ public interface PlayerDao {
 	
 	Player findPlayerByName(String name);
 	
-	List<Player> findPlayerByRoomId(int roomId);
+	List<Player> findPlayerByRoomId(long roomId);
 }
